+++ src/App.vue
@@ -13,7 +13,7 @@
 // import AskMain from "@/components/franchise/ask/AskMain.vue";
 // import Breadcrumb from "@/components/franchise/ask/Breadcrumb.vue";
 // import CategoryList from "@/components/amdin/Category/CategoryList.vue";
-import ProductList11 from "@/components/amdin/product/ProductList11.vue";
+import ProductList11 from "@/components/amdin/product/ProductList.vue";
 
 </script>
